sedes

+++ resources/views/archivos/paquetes/index.blade.php
@@ -53,11 +53,13 @@
 
 								<td>
 									<a href="#" class="btn btn-primary view" onclick="view(this)" data-id="{{$paq->id}}">ver</a>
+								@if(\Auth::user()->role_id <> 6)
 									<a href="paquetes-edit-{{$paq->id}}" class="btn btn-warning edit">Editar</a>
 									<a href="paquetes-delete-{{$paq->id}}" class="btn btn-danger">Eliminar</a>
 									@if ($paq->id != 1)
 										<a href="paquetes-addItems-{{$paq->id}}" class="btn btn-success">Agregar items</a>
 									@endif
+								@endif	
 								</td>
 							</tr>
 						@endforeach
